src/ocr/line_detect.py
"""Line detector wrapper: CRAFT/DBNet-style detection skeleton."""
import cv2
import logging

def detect_lines(image):
    try:
        h, w = image.shape[:2]
        logging.info(f"Running line detection on image of size {w}x{h}")
        step = max(32, h // 10)
        out = []
        for y in range(0, h, step):
            out.append({
                "label": "TextLine",
                "bbox": [0, y, w, min(step, h - y)],
                "score": 0.99,
                "text": f"Line at y={y}",
                "lang": "en"
            })
        logging.info(f"Detected {len(out)} text lines.")
        return out
    except Exception as e:
        logging.error(f"Error in line detection: {e}")
        return []
if __name__ == '__main__':
    import numpy as np
    img = 255*np.ones((800,600,3), dtype='uint8')
    print('Detections:', detect_lines(img))
